Support using custom source and target dir

+++ main.py
@@ -1,23 +1,33 @@
+import sys
 import os
-from os.path import join as pathjoin
 from os import mkdir as mkdir
+from os import path as p
 
-import fileIO
 from htmlobj import *
-from fileIO import fopen as open
 from pages import *
+from cmdio import *
+
+def printHelp():
+    logClean("usage:")
+    logClean(sys.argv[0] + " -t / --target [target_dir]")
+    logClean(sys.argv[0] + " -s / --source [source_dir]")
+    logClean(sys.argv[0] + " -h / --help   print this information")
 
 def genTargetDirs(targetdir: str) -> None:
     if os.path.exists(targetdir):
-         os.system("rm -r "+targetdir)
+        if askYesOrNo("Target directory exists, delete? (y/N):", default=False):
+            os.system("rm -r "+targetdir)
+        else:
+            exit(2)
+
     mkdir(targetdir)
-    targetresdir = pathjoin(targetdir, "res/")
+    targetresdir = p.join(targetdir, "res/")
     mkdir(targetresdir)
     
-    print("copying resources...", end="")
+    log("copying resources...", end="")
     resdir = os.path.abspath("res/")
     os.system("cp -rf "+resdir+"/* "+targetdir)
-    print("ok")
+    logClean("ok")
 
 def loadArticlesInDir(articlesDir: str) -> list[Article]:
     articles = []
@@ -27,29 +37,50 @@ def loadArticlesInDir(articlesDir: str) -> list[Article]:
         fileext = lol[1]
         if fileext != ".json":
             continue
-        path = pathjoin(articlesDir, filename)
-        print("loading "+path+"...", end="")
+        path = p.join(articlesDir, filename)
+        log("loading "+path+"...", end="")
         articles.append(Article(path))
-        print("ok")
+        logClean("ok")
     return articles
 
 def main():
     targetdir = os.path.abspath("generated/")
+    srcdir = os.path.abspath(".")
+
+    i = 1
+    while i < len(sys.argv):
+        arg = sys.argv[i]
+        argL = arg.lower()
+        if argL == "-t" or argL == "--target":
+            targetdir = sys.argv[i+1]
+            i += 1
+        elif argL == "-s" or argL == "--source":
+            srcdir = sys.argv[i+1]
+            i+= 1
+        elif argL == "-h" or argL == "-h":
+            printHelp()
+            exit(0)
+        else:
+            logErr("Unknown argument: "+arg)
+            printHelp()
+            exit(3)
+        i += 1
+
     genTargetDirs(targetdir)
-    postsdir = pathjoin(targetdir, "posts/")
+    postsdir = p.join(targetdir, "posts/")
     mkdir(postsdir)
 
-    homePage = HomePage("templates/home/index", "templates/home/blogitem")
+    homePage = HomePage(p.join(srcdir, "templates/home/index"), p.join(srcdir, "templates/home/blogitem"))
 
-    for article in loadArticlesInDir("articles"):
+    for article in loadArticlesInDir(path.join(srcdir, "articles")):
         article.saveToDir(postsdir)
         homePage.addSubpageOfArticle(article)
-    print("all articles generated")
-    print("generating homepage...", end="")
-    homePage.writeToFile(pathjoin(targetdir, "index.html"))
-    print("ok")
+    log("all articles generated")
+    log("generating homepage...", end="")
+    homePage.writeToFile(p.join(targetdir, "index.html"))
+    logClean("ok")
 
-    print("all generated, wrapping up...")
+    log("all generated, wrapping up...")
     fileIO.closeAllOpenedFiles()
 
 main()
